ID-1025 Update product fragments and queries
