--- Core/GraphQL/GraphQLFiles/sell_fragment_products.graphql
@@ -1,6 +1,0 @@
-fragment FragmentProducts on ProductEdge {
-  edges {
-    ...FragmentProduct
-  }
-  nextPage
-}
