+++ Core/GraphQL/GraphQLFiles/sell_getProducts.graphql
@@ -1,5 +1,8 @@
 query getProducts($pageSize: Int!, $page: Int, $filters: ProductFilter) {
   getProducts(pageSize: $pageSize, page: $page, filters: $filters) {
-    ...FragmentProducts
+    edges {
+      ...FragmentProduct
+    }
+    nextPage
   }
 }
