+++ Core/GraphQL/schema.json
@@ -4408,7 +4408,7 @@
                   "name": null,
                   "ofType": {
                     "kind": "INPUT_OBJECT",
-                    "name": "PaymentIntentInput",
+                    "name": "PaymentIntentUpdateInput",
                     "ofType": null
                   }
                 },
@@ -10065,20 +10065,6 @@
             "defaultValue": null
           },
           {
-            "name": "paymentCustomer",
-            "description": null,
-            "type": {
-              "kind": "NON_NULL",
-              "name": null,
-              "ofType": {
-                "kind": "INPUT_OBJECT",
-                "name": "PaymentCustomerInput",
-                "ofType": null
-              }
-            },
-            "defaultValue": null
-          },
-          {
             "name": "amount",
             "description": null,
             "type": {
@@ -10150,6 +10136,47 @@
             "type": {
               "kind": "SCALAR",
               "name": "String",
+              "ofType": null
+            },
+            "defaultValue": null
+          }
+        ],
+        "interfaces": null,
+        "enumValues": null,
+        "possibleTypes": null
+      },
+      {
+        "kind": "INPUT_OBJECT",
+        "name": "PaymentIntentUpdateInput",
+        "description": null,
+        "fields": null,
+        "inputFields": [
+          {
+            "name": "status",
+            "description": null,
+            "type": {
+              "kind": "ENUM",
+              "name": "UpdatePaymentStatus",
+              "ofType": null
+            },
+            "defaultValue": null
+          },
+          {
+            "name": "paymentSource",
+            "description": null,
+            "type": {
+              "kind": "INPUT_OBJECT",
+              "name": "PaymentSourceInput",
+              "ofType": null
+            },
+            "defaultValue": null
+          },
+          {
+            "name": "savePaymentSource",
+            "description": null,
+            "type": {
+              "kind": "SCALAR",
+              "name": "Boolean",
               "ofType": null
             },
             "defaultValue": null
@@ -10507,6 +10534,23 @@
             "isDeprecated": false,
             "deprecationReason": null
           },
+          {
+            "name": "canceled",
+            "description": null,
+            "isDeprecated": false,
+            "deprecationReason": null
+          }
+        ],
+        "possibleTypes": null
+      },
+      {
+        "kind": "ENUM",
+        "name": "UpdatePaymentStatus",
+        "description": null,
+        "fields": null,
+        "inputFields": null,
+        "interfaces": null,
+        "enumValues": [
           {
             "name": "canceled",
             "description": null,
